requirements.txt
# Core Dependencies
python>=3.10

# Microsoft Azure AI & Agent Framework
agent-framework-azure-ai>=0.1.0

# Web Scraping & Parsing
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Document Processing
unstructured>=0.11.0
pypdf>=3.17.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0

# Vector Store & Embeddings
agentset-sdk>=1.0.0
chromadb>=0.4.0

# Data Storage
boto3>=1.33.0
sqlalchemy>=2.0.0

# Scheduling & Orchestration
apache-airflow>=2.7.0
celery>=5.3.0

# Validation & Testing
pytest>=7.4.0
great-expectations>=0.18.0
pytest-asyncio>=0.21.0

# Monitoring & Metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
asyncio>=3.4.3
aiohttp>=3.9.0
tenacity>=8.2.0
loguru>=0.7.0
